fixed miner lookup by coinbase addr duplicate bug, using a map now instead of list, to still account for remote posibility of multiple miners depositing to same address. fixed Transaction value overflow bug, where int (32bit default) is used where values are much larger and overflow; now using uint64. same for gas prices and tx cost.

+++ go/gf_eth_monitor_core/gf_eth_tx.go
@@ -26,15 +26,15 @@ import (
 //-------------------------------------------------
 type GF_eth__tx struct {
 	Hash_str      string           `json:"hash_str"      bson:"hash_str"`
-	Index_int     uint             `json:"index_int"     bson:"index_int"` // position of the transaction in the block
+	Index_int     uint64           `json:"index_int"     bson:"index_int"` // position of the transaction in the block
 	From_addr_str string           `json:"from_addr_str" bson:"from_addr_str"`
 	To_addr_str   string           `json:"to_addr_str"   bson:"to_addr_str"`
-	Value_int     int64            `json:"value_int"     bson:"value_int"`
+	Value_int     uint64           `json:"value_int"     bson:"value_int"`
 	Gas_used_int  uint64           `json:"gas_used_int"  bson:"gas_used_int"`
-	Gas_price_int int64            `json:"gas_price_int" bson:"gas_price_int"`
+	Gas_price_int uint64           `json:"gas_price_int" bson:"gas_price_int"`
 	Nonce_int     uint64           `json:"nonce_int"     bson:"nonce_int"`
 	Size_f        float64          `json:"size_f"        bson:"size_f"`
-	Cost_int      int64            `json:"cost_int"      bson:"cost_int"`
+	Cost_int      uint64           `json:"cost_int"      bson:"cost_int"`
 	Logs          []*eth_types.Log `json:"logs_lst"      bson:"logs_lst"`
 }
